Updated components to use bv inheritance structure (boundingvolume.h)
