+++ Game/GameMain_Test/test.cpp
@@ -1,9 +1,9 @@
 #include "gtest/gtest.h"
 #include <crtdbg.h>
 
-//#include "Physics/GroundCollision.h"
-//#include "Physics/MovementLogic.h"
-//#include "Physics/ObjectCollision.h"
+#include "Physics/GroundCollision.h"
+#include "Physics/MovementLogic.h"
+#include "Physics/ObjectCollision.h"
 #include "Physics/FightingLogic.h"
 //#include "Physics/QuadTreeTests.h"
 //#include "Sound/SoundTests.h"
